Create "Status Report: The Funkin’ Mobile Trailer and Plushies!"
My desktop users, there is no need for scrolling for this post, I've configured properly at the beginning for once! As for my mobile users, still working on that, might need to inject some actual code into this to make it work... but unless status update for Funkin' cause a lot happened this week and I want to share it with y'all!